resovle comment 80%
