+++ backend/internals/entities/payload/strengthData_response.go
@@ -1,0 +1,6 @@
+package payload
+
+type StrengthDataResponse struct {
+	FieldName string `json:"fieldName"`
+	TotalGems int64  `json:"totalGems"`
+}
