--- backend/internals/entities/response/strengthDataResponse.go
@@ -1,6 +1,0 @@
-package response
-
-type StrengthDataResponse struct {
-	FieldName string `json:"field_name"`
-	TotalGems int64  `json:"total_gems"`
-}
