+++ backend/internals/repositories/enroll.go
@@ -1,5 +1,9 @@
 package repositories
 
-type EnrollRepo interface {
-	EnrollUser(userId uint, courseId uint64) error
+import "backend/internals/db/models"
+
+type EnrollRepository interface {
+	FindEnrollmentsByUserID(userId *string) ([]models.Enroll, error)
+	GetTotalStepsByCourseID(courseId uint64) (int64, error)
+	GetEvaluatedStepsByUserAndCourse(userId uint64, courseId uint64) (int64, error)
 }
