+++ backend/internals/repositories/userActivity.go
@@ -1,0 +1,7 @@
+package repositories
+
+import "backend/internals/db/models"
+
+type UserActivityRepository interface {
+	GetRecentActivitiesByUserID(userId *string) ([]models.UserActivity, error)
+}
