+++ backend/internals/repositories/userStrength.go
@@ -1,0 +1,11 @@
+package repositories
+
+import (
+	"backend/internals/db/models"
+	"backend/internals/entities/payload"
+)
+
+type UserStrengthRepository interface {
+	GetStrengthDataByUserID(userId uint64) ([]payload.StrengthDataResponse, error)
+	GetSuggestionCourse(userId uint64) ([]models.Course, error)
+}
