+++ backend/internals/services/enroll.go
@@ -1,5 +1,8 @@
 package services
 
-type EnrollServices interface {
+import "backend/internals/entities/payload"
+
+type EnrollService interface {
+	GetEnrollmentsByUserID(userId *string) (payload.EnrollmentListResponse, error)
 	EnrollUser(userId uint, courseId uint64) error
 }
