+++ backend/internals/services/userActivity.go
@@ -1,0 +1,7 @@
+package services
+
+import "backend/internals/entities/payload"
+
+type UserActivityService interface {
+	GetRecentActivitiesByUserID(userId *string) (*payload.UserActivitiesResponse, error)
+}
