+++ frontend/src/components/ui/SuggestionCard.tsx
@@ -1,4 +1,3 @@
-// Update SuggestionCard component
 import { Card, CardContent } from "@/components/ui/card"
 import { BookOpen } from "lucide-react"
 
@@ -9,14 +8,14 @@ export interface SuggestionCardProps {
     }
 }
 
-export const SuggestionCard = ({ name = "Untitled Course", field = { name: "INFRASTRUCTURE" } }: SuggestionCardProps) => (
+export const SuggestionCard = ({ name, field }: SuggestionCardProps) => (
     <Card className="w-full h-full">
         <CardContent className="p-6">
             <div className="flex flex-col">
                 <div className="flex items-center gap-2 mb-3">
                     <span className="text-gray-500 text-sm font-medium inline-flex items-center">
                         <BookOpen className="w-4 h-4 mr-1" />
-                        COURSE / {field.name}
+                        COURSE / {field?.name || "GENERAL"}
                     </span>
                 </div>
 
@@ -25,9 +24,14 @@ export const SuggestionCard = ({ name = "Untitled Course", field = { name: "INFR
                 </div>
 
                 <div className="mt-auto">
-                    <img src="/src/assets/microcontroller.png" alt="Course icon" className="w-8 h-8 object-contain opacity-80" />
+                    <img
+                        src="https://static.bookmark.scnd.app/asset/fieldicon/microcontroller.png"
+                        alt="Course icon"
+                        className="w-12 h-12 object-contain opacity-80"
+                    />
                 </div>
             </div>
         </CardContent>
     </Card>
 )
+
